modularize flattenData()
